app/requirements.txt
numpy
pandas
pillow
streamlit
tensorflow
